feat: 캡슐화 강의 정리
