+++ lib/240311/00_encapsulation/king.dart
@@ -1,0 +1,10 @@
+
+import 'hero.dart';
+
+class King {
+
+  void callHero(Hero hero) {
+    print('용사님, 저희 왕국에 와주셔서 감사합니다.');
+    print('용사님의 이름은 ${hero.name}이고, HP는 ${hero.hp}입니다.');
+  }
+}
