+++ lib/240311/00_encapsulation/slime.dart
